requirements.txt
amqp==5.1.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
billiard==3.6.4.0
blessed==1.19.1
blinker==1.4
bpython==0.24
Brlapi==0.8.3
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
curtsies==0.4.1
cwcwidth==0.1.8
cycler==0.11.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
duplicity==0.8.21
executing==1.2.0
fasteners==0.14.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
greenlet==2.0.2
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
ipython==8.10.0
itypes==1.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
Pygments==2.15.0
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-env==1.0.0
pythran==0.10.0
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
redis==4.5.1
reportlab==3.6.8
requests==2.28.2
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.9
systemd-python==234
terminator==2.1.1
traitlets==5.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.15
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
